# api/openapi/sanya-spb-URL-shortener-1.1.0-swagger.json
{
  "openapi" : "3.0.0",
  "servers" : [ ],
  "info" : {
    "description" : "This is a simple API for GB/Go Backend level1",
    "version" : "1.1.0",
    "title" : "URL-shortener",
    "contact" : {
      "email" : "sanya-spb@mail.ru"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "tags" : [ {
    "name" : "admin",
    "description" : "Secured Admin-only calls"
  }, {
    "name" : "user",
    "description" : "Operations available to regular users"
  } ],
  "paths" : {
    "/" : {
      "put" : {
        "tags" : [ "user" ],
        "summary" : "adds Link",
        "operationId" : "addLink",
        "description" : "method Create (in CRUD spec)",
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Link"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "409" : {
            "description" : "Conflict"
          },
          "500" : {
            "description" : "Internal Server Error"
          },
          "501" : {
            "description" : "Not Implemented"
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : [ "Name", "URL", "Descr" ],
                "properties" : {
                  "Name" : {
                    "type" : "string",
                    "example" : "Google"
                  },
                  "URL" : {
                    "type" : "string",
                    "format" : "url",
                    "example" : "https://google.com/index.html"
                  },
                  "Descr" : {
                    "type" : "string",
                    "example" : "Google main page v.1"
                  },
                  "DeleteAt" : {
                    "type" : "string",
                    "format" : "date-time",
                    "example" : "2022-01-01T11:11:11.111Z"
                  }
                }
              }
            }
          },
          "description" : "Link info to add"
        }
      }
    },
    "/{uuid}" : {
      "put" : {
        "tags" : [ "admin" ],
        "summary" : "updates Link",
        "operationId" : "updLink",
        "description" : "method Update (in CRUD spec)",
        "parameters" : [ {
          "in" : "path",
          "name" : "uuid",
          "description" : "pass an uuid for updating",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "409" : {
            "description" : "Conflict"
          },
          "500" : {
            "description" : "Internal Server Error"
          },
          "501" : {
            "description" : "Not Implemented"
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : [ "Name", "URL", "Descr" ],
                "properties" : {
                  "Name" : {
                    "type" : "string",
                    "example" : "Google"
                  },
                  "URL" : {
                    "type" : "string",
                    "format" : "url",
                    "example" : "https://google.com/index.html"
                  },
                  "Descr" : {
                    "type" : "string",
                    "example" : "Google main page v.2"
                  },
                  "DeleteAt" : {
                    "type" : "string",
                    "format" : "date-time",
                    "example" : "2022-02-02T22:22:22.222Z"
                  }
                }
              }
            }
          },
          "description" : "Link info to update"
        }
      }
    },
    "/r/{uuid}" : {
      "post" : {
        "tags" : [ "admin" ],
        "summary" : "updates Link and return updated Link",
        "operationId" : "updRetLink",
        "description" : "method Update (in CRUD spec)",
        "parameters" : [ {
          "in" : "path",
          "name" : "uuid",
          "description" : "pass an uuid for updating",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Link"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "409" : {
            "description" : "Conflict"
          },
          "500" : {
            "description" : "Internal Server Error"
          },
          "501" : {
            "description" : "Not Implemented"
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : [ "Name", "URL", "Descr" ],
                "properties" : {
                  "Name" : {
                    "type" : "string",
                    "example" : "Google"
                  },
                  "URL" : {
                    "type" : "string",
                    "format" : "url",
                    "example" : "https://google.com/index.html"
                  },
                  "Descr" : {
                    "type" : "string",
                    "example" : "Google main page v.2"
                  },
                  "DeleteAt" : {
                    "type" : "string",
                    "format" : "date-time",
                    "example" : "2022-02-02T22:22:22.222Z"
                  }
                }
              }
            }
          },
          "description" : "Link info to update"
        }
      }
    },
    "/info/{uuid}" : {
      "get" : {
        "tags" : [ "user" ],
        "summary" : "gets Link info",
        "operationId" : "getLink",
        "description" : "method READ (in CRUD spec)",
        "parameters" : [ {
          "in" : "path",
          "name" : "uuid",
          "description" : "pass an uuid for looking up Link",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Link"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Internal Server Error"
          },
          "501" : {
            "description" : "Not Implemented"
          }
        }
      }
    },
    "/delete/{uuid}" : {
      "delete" : {
        "tags" : [ "admin" ],
        "summary" : "delete Link",
        "operationId" : "delLink",
        "description" : "method Delete (in CRUD spec)",
        "parameters" : [ {
          "in" : "path",
          "name" : "uuid",
          "description" : "pass an uuid to delete",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Internal Server Error"
          },
          "501" : {
            "description" : "Not Implemented"
          }
        }
      }
    },
    "/delete/r/{uuid}" : {
      "delete" : {
        "tags" : [ "admin" ],
        "summary" : "delete Link and return deleted Link",
        "operationId" : "delRetLink",
        "description" : "method Delete (in CRUD spec)",
        "parameters" : [ {
          "in" : "path",
          "name" : "uuid",
          "description" : "pass an uuid to delete",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Link"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Internal Server Error"
          },
          "501" : {
            "description" : "Not Implemented"
          }
        }
      }
    },
    "/{id}" : {
      "get" : {
        "tags" : [ "user" ],
        "summary" : "go to source link",
        "operationId" : "goTo",
        "description" : "By passing in the appropriate options, you can move external www source\n",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "description" : "pass an short link as id to move",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "302" : {
            "description" : "Found"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Internal Server Error"
          },
          "501" : {
            "description" : "Not Implemented"
          }
        }
      }
    },
    "/stat" : {
      "get" : {
        "tags" : [ "admin" ],
        "summary" : "get statistic",
        "operationId" : "getStat",
        "description" : "By passing in the appropriate options, you can get stats for using URL-shorterner\n",
        "parameters" : [ {
          "in" : "query",
          "name" : "q",
          "description" : "pass an optional search string for looking up inventory",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "found, see stat",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Stat"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Internal Server Error"
          },
          "501" : {
            "description" : "Not Implemented"
          }
        }
      }
    },
    "/status" : {
      "get" : {
        "tags" : [ "user" ],
        "summary" : "get status",
        "operationId" : "getStatus",
        "description" : "always return ok",
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string",
                  "example" : "ok"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error"
          },
          "501" : {
            "description" : "Not Implemented"
          }
        }
      }
    },
    "/status/{code}" : {
      "get" : {
        "tags" : [ "admin" ],
        "summary" : "echo code",
        "operationId" : "echoStatus",
        "description" : "always echo code",
        "parameters" : [ {
          "in" : "path",
          "name" : "code",
          "description" : "pass an uuid for looking up Link",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string",
                  "example" : "codeOK"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error"
          },
          "501" : {
            "description" : "Not Implemented"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "Link" : {
        "type" : "object",
        "required" : [ "UUID", "URL", "Name", "Descr", "LinkID", "CreatedAt", "DeleteAt", "User" ],
        "properties" : {
          "UUID" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "123e4567-e89b-12d3-a456-426655440000"
          },
          "URL" : {
            "type" : "string",
            "format" : "url",
            "example" : "https://google.com/index.html"
          },
          "Name" : {
            "type" : "string",
            "example" : "Google"
          },
          "Descr" : {
            "type" : "string",
            "example" : "Google main page"
          },
          "LinkID" : {
            "type" : "string",
            "example" : "wohv7Boh"
          },
          "CreatedAt" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2031-08-29T09:12:33.001Z"
          },
          "DeleteAt" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2031-08-29T09:12:33.001Z"
          },
          "User" : {
            "$ref" : "#/components/schemas/User"
          }
        }
      },
      "User" : {
        "required" : [ "nickname" ],
        "properties" : {
          "nickname" : {
            "type" : "string",
            "example" : "James Fenimore Cooper"
          }
        }
      },
      "Stat" : {
        "required" : [ "link", "redirects" ],
        "properties" : {
          "link" : {
            "$ref" : "#/components/schemas/Link"
          },
          "redirects" : {
            "type" : "integer",
            "format" : "uint64",
            "example" : 65536
          }
        },
        "type" : "array",
        "items" : {
          "allOf" : [ {
            "$ref" : "#/components/schemas/Link"
          }, {
            "type" : "integer"
          } ]
        }
      }
    }
  }
}